requirements.txt
docopt==0.6.2
PyYAML==3.11
requests==2.7.0
wheel==0.24.0
